feat: restructure components to use new _components directory
