+++ app/(client)/home/loading.tsx
@@ -1,4 +1,4 @@
-import { LoadingSkeleton } from '../../components/loading-skeleton'
+import { LoadingSkeleton } from '../../_components/loading-skeleton'
 
 export default function Loading() {
   return (
